--- src/screens/KakaoLoginView.tsx
@@ -1,55 +1,0 @@
-import React from 'react';
-import useAuth from '@hooks/useAuth';
-import { setLogin } from '@api/auth/auth';
-import WebView from 'react-native-webview';
-import { SocialLoginProps } from '@type/index';
-import { SafeAreaView } from 'react-native-safe-area-context';
-import { REACT_APP_REST_API_KEY, REACT_APP_REDIRECT_URI } from '@env';
-import { Platform } from 'react-native';
-
-export default function KakaoLoginWebView() {
-  const INJECTED_JAVASCRIPT = `window.ReactNativeWebView.postMessage("check")`;
-  const { setLogin: login } = useAuth();
-
-  function LogInProgress(data: string) {
-    const exp = 'code=';
-    const condition = data.indexOf(exp);
-
-    if (condition !== -1) {
-      const requestCode = data.substring(condition + exp.length);
-
-      requestToken(requestCode);
-    }
-  }
-
-  const requestToken = async (requestCode: string) => {
-    const requestTokenUrl = '/api/v1/oauth/kakao';
-    const data: SocialLoginProps = {
-      url: requestTokenUrl,
-      data: { code: requestCode },
-    };
-
-    setLogin(data).then(res => login(res));
-  };
-
-  return (
-    <SafeAreaView style={{ flex: 1, width: '100%' }}>
-      <WebView
-        originWhitelist={[`*`]}
-        scalesPageToFit
-        style={{ marginTop: 30 }}
-        source={{
-          uri: `https://kauth.kakao.com/oauth/authorize?client_id=${REACT_APP_REST_API_KEY}&redirect_uri=${REACT_APP_REDIRECT_URI}&response_type=code&`,
-        }}
-        injectedJavaScript={INJECTED_JAVASCRIPT}
-        onMessage={event => {
-          if (Platform.OS === 'android' && !event.nativeEvent.loading) {
-            LogInProgress(event.nativeEvent.url);
-          } else if (Platform.OS === 'ios' && event.nativeEvent.loading) {
-            LogInProgress(event.nativeEvent.url);
-          }
-        }}
-      />
-    </SafeAreaView>
-  );
-}
